requirements.txt
fastapi[all]
uvicorn
asyncio
gunicorn[gthread]
numpy
pandas==2.2.2
Requests
aiohttp
multiprocess
scikit_learn
python-dotenv